store: add oscapEndpoints for composerApi
