+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -1,2 +1,3 @@
+export { lookupDatastreamDistro } from './dataStreamLookup';
 export { getBlueprintsPath } from './getBlueprintsPath';
 export { getCloudConfigs } from './getCloudConfigs';
